fix(mobile): firmware update dev settings crash in production
FirmwareUpdateGalleryDemo was imported from Developer/pages/Gallery,
which the Metro resolver replaces with an empty stub when UNION_BUILD
or SPLIT_BUNDLE_SEGMENTS is enabled. The named export resolved to
undefined at runtime, triggering "Element type is invalid".

Move the demo helpers out of Gallery into DevSettingsSection as
FirmwareUpdateActions, and drop the now-unused Gallery route entry.
