+++ packages/kit/src/views/Developer/pages/Gallery/index.tsx
@@ -73,11 +73,6 @@ const ErrorToastGallery = LazyLoadPage(
     import('@onekeyhq/kit/src/views/Developer/pages/Gallery/Components/stories/ErrorToastGallery'),
 );
 
-const FirmwareUpdateGallery = LazyLoadPage(
-  () =>
-    import('@onekeyhq/kit/src/views/Developer/pages/Gallery/Components/stories/FirmwareUpdateGallery'),
-);
-
 const QRWalletGallery = LazyLoadPage(
   () =>
     import('@onekeyhq/kit/src/views/Developer/pages/Gallery/Components/stories/QRWalletGallery'),
@@ -631,10 +626,6 @@ export const galleryScreenList: {
   { name: EGalleryRoutes.ComponentJotaiGlobal, component: JotaiGlobalGallery },
   { name: EGalleryRoutes.ComponentLocalDB, component: LocalDBGallery },
   { name: EGalleryRoutes.ComponentErrorToast, component: ErrorToastGallery },
-  {
-    name: EGalleryRoutes.ComponentFirmwareUpdate,
-    component: FirmwareUpdateGallery,
-  },
   {
     name: EGalleryRoutes.ComponentQRWallet,
     component: QRWalletGallery,
